Fixed incorrect assumption that pid_t is int.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@36071 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/fsync/4-1.c
@@ -34,8 +34,8 @@ int main()
   int total_size = 1024; 
   int fd;
 
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_fsync_4_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_fsync_4_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
