+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/31-1.c
@@ -52,8 +52,8 @@ int main()
   long page_size = sysconf(_SC_PAGE_SIZE);
 
   
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_31_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_31_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
