+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/5-1.c
@@ -48,8 +48,8 @@ int main()
   int prot;
   off_t off = 0;
 
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_5_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_5_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
