+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/sem_wait/7-1.c
@@ -39,7 +39,7 @@ int main()
         char semname[20];
         int pid, status;
 
-	sprintf(semname, "/" FUNCTION "_" TEST "_%d", getpid());
+	sprintf(semname, "/" FUNCTION "_" TEST "_%ld", (long)getpid());
 
         mysemp = sem_open(semname, O_CREAT, 0, 1);
         if( mysemp == SEM_FAILED || mysemp == NULL ) {
